V102 (#16)
* Strip IPs covered by subnets.

* Code tuning.

* Variable name tuning.

* Code tuning.

* Code linting.

* Code tuning.

* Fixed #13
The implementation of the optimized binary search
algorithm was not exactly as written in the issue.
Upon reading the python documentation more
closely, it's appropriate to treat IP Network
objects like addressable lists (using []
notation). Doing so significantly improved the
algorithm's performance.

* Fixed #14

* Variable name tuning.

* Fixed #15

* Acknowledgements
Added license and acknowledgements for the tqmd
library, and bumped tqmd to v4.66.5